Simplify configuration for Capella
Take advantage of the Capella CA certificate bundled with Java SDK 3.3.0.

Add a note that the simplest way to enable TLS is to treat the
`spring.couchbase.bootstrap-hosts` application property as a
connection string, and prefix it with "couchbases://".

In the commented-out example code, configure the ClusterEnvironment
using a lambda. This way, the user is not responsible for shutting down
the environment.

Disconnect the cluster when the bean is destroyed. It doesn't matter
for this example project, but it's a good practice that prevents
resource leakage when the Spring webapp is undeployed from a
non-embedded app server.